moved barAdder to components
