+++ js/components/barAdder.js
